+++ app/controllers/api/v1/book_controller.rb
@@ -1,6 +1,6 @@
 module Api
   module V1
-    class BookController < ApiController
+    class BookController < Api::V1::ApiController
       def index
         @books = Book.all
         render_paginated json: @books
